fix(tests): move mock geolocator and revert profile page
