+++ integration_test/app_test.dart
@@ -20,6 +20,7 @@ import "package:proxima/views/pages/new_post/new_post_form.dart";
 import "package:proxima/views/pages/profile/profile_page.dart";
 
 import "../test/mocks/mock_geo_location_service.dart";
+import "../test/mocks/overrides/override_user_repo.dart";
 import "../test/mocks/setup_firebase_mocks.dart";
 import "../test/services/firebase/testing_auth_providers.dart";
 
@@ -52,7 +53,7 @@ void main() {
       ProviderScope(
         overrides: [
           ...firebaseAuthMocksOverrides,
-          userRepositoryProvider.overrideWithValue(userRepo),
+          userRepoOverride,
           geoLocationServiceProvider.overrideWithValue(geoLocationService),
           postRepositoryProvider.overrideWithValue(postRepo),
         ],
