+++ test/views/pages/profile/profile_page_test.dart
@@ -4,6 +4,8 @@ import "package:flutter/material.dart";
 import "package:flutter_test/flutter_test.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/models/database/user/user_firestore.dart";
+import "package:proxima/services/database/user_repository_service.dart";
+import "package:proxima/services/login_service.dart";
 import "package:proxima/views/navigation/routes.dart";
 import "package:proxima/views/pages/home/top_bar/app_top_bar.dart";
 import "package:proxima/views/pages/profile/posts_info/info_card_badge.dart";
@@ -16,7 +18,6 @@ import "package:proxima/views/pages/profile/profile_page.dart";
 import "package:proxima/views/pages/profile/user_info/user_account.dart";
 
 import "../../../mocks/data/mock_firestore_user.dart";
-import "../../../mocks/overrides/override_user_repo.dart";
 import "../../../mocks/pages/mock_homepage.dart";
 import "../../../mocks/setup_firebase_mocks.dart";
 import "../../../services/firebase/testing_auth_providers.dart";
@@ -24,6 +25,7 @@ import "../../../services/firebase/testing_auth_providers.dart";
 void main() {
   late FakeFirebaseFirestore fakeFireStore;
   late CollectionReference<Map<String, dynamic>> userCollection;
+  late UserRepositoryService userRepo;
   late ProviderScope mockedProfilePage;
 
   final expectedUser = testingUserFirestore;
@@ -32,15 +34,17 @@ void main() {
     setupFirebaseAuthMocks();
     fakeFireStore = FakeFirebaseFirestore();
     userCollection = fakeFireStore.collection(UserFirestore.collectionName);
-
+    userRepo = UserRepositoryService(
+      firestore: fakeFireStore,
+    );
     await userCollection
         .doc(expectedUser.uid.value)
         .set(expectedUser.data.toDbData());
 
     mockedProfilePage = ProviderScope(
       overrides: [
-        ...firebaseAuthMocksOverrides,
-        userRepoOverride,
+        firebaseAuthProvider.overrideWith(mockFirebaseAuthSignedIn),
+        userRepositoryProvider.overrideWithValue(userRepo),
       ],
       child: const MaterialApp(
         onGenerateRoute: generateRoute,
